Copy structure from DALA-7363
