+++ dataland-backend/src/main/kotlin/org/dataland/datalandbackend/services/datapoints/DataPointConversionComments.kt
@@ -2,6 +2,7 @@ package org.dataland.datalandbackend.services.datapoints
 
 import org.dataland.datalandbackend.model.datapoints.UploadedDataPoint
 import org.dataland.datalandbackend.model.enums.data.QualityOptions
+import org.dataland.datalandbackend.model.enums.eutaxonomy.nonfinancials.Activity
 import org.dataland.datalandbackendutils.model.DataPointType
 import org.dataland.specificationservice.openApiClient.model.DataPointTypeSpecification
 import org.dataland.specificationservice.openApiClient.model.FrameworkSpecification
@@ -85,3 +86,45 @@ internal fun getSourceFrameworkLabel(sourceFrameworks: List<FrameworkSpecificati
         .takeIf { it.isNotEmpty() }
         ?.joinToString(", ")
         ?: "Unknown"
+
+/**
+ * Extends [baseComment] with notes about activities affected by the EU taxonomy activity merge.
+ *
+ * @param baseComment the comment generated for the merge conversion before these notes are appended
+ * @param activitiesWithConflictingSubstantialContributions activities for which conflicting substantial
+ *   contributions were detected and removed
+ * @param activitiesWithoutAlignedShares activities for which the relative aligned share was null, so their
+ *   substantial contributions were removed
+ * @return [baseComment] extended with the applicable notes
+ */
+internal fun extendCreateCommentEuTaxonomyActivitiesMerge(
+    baseComment: String,
+    activitiesWithConflictingSubstantialContributions: List<Activity>,
+    activitiesWithoutAlignedShares: List<Activity>,
+): String {
+    var comment = baseComment
+    comment =
+        if (activitiesWithConflictingSubstantialContributions.isNotEmpty()) {
+            comment + "\n\n" +
+                "For the following activities, more than one substantial contribution was reported as non-zero " +
+                "in the old framework. Since the substantial contributions are expected to sum up to the " +
+                "activity's aligned relative share, but the correct scaling could not be determined from the " +
+                "old framework, no substantial contributions were mapped for these activities:\n\n" +
+                activitiesWithConflictingSubstantialContributions.joinToString("\n") { "- ${it.value}" }
+        } else {
+            comment
+        }
+
+    comment =
+        if (activitiesWithoutAlignedShares.isNotEmpty()) {
+            comment + "\n\n" +
+                "Activities without relative aligned share cannot have substantial contributions," +
+                "whose values must add up to the relative share. Thus, the substantial contributions" +
+                "were removed for these activities:\n\n" +
+                activitiesWithoutAlignedShares.joinToString("\n") { "- ${it.value}" }
+        } else {
+            comment
+        }
+
+    return comment
+}
